Updates in Rest

+++ Capgemini Spring Rest Assignment/Assignments_1/question_7/src/main/java/com/example/question_7/Repository/OrderRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderRepository extends MongoRepository<Order,Integer> {
+public interface OrderRepository extends MongoRepository<Order,Long> {
 
 }
